chore: renovate backend controller setup, new schema addition
