+++ server/controllers/controller.ts
@@ -1,0 +1,3 @@
+export * from './userController'
+export * from './tripController'
+export * from './activityController'
